requirements.txt
azure-common==1.1.28
azure-core==1.26.0
azure-identity==1.11.0
azure-keyvault-secrets==4.6.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
clint==0.5.1
colorama==0.4.5
cryptography==38.0.1
idna==3.4
isodate==0.6.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.6.0
pycparser==2.21
pyfiglet==0.8.post1
PyJWT==2.6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
tabulate==0.9.0
typing_extensions==4.4.0
urllib3==1.26.12
